Assignment helper refactor (#2139)
* Helper & Model Changes

* Include with ActionView

* Create scoring.rb in lib\

* Change autoload_paths

* Eager_load_paths addition

* spec changes - allow(Scoring)...

* New spec allow() attempt

* Additional spec changes, travis lib addition test

* Scoring_spec test method call fix

* Scoring spec update

* Extend scoring inside response_map

Extending it makes the methods in Scoring be a method on the class itself, instead of an instance method that must be called on an object

* Try allow_any_instance_of(Scoring)

* remove unused helper methods

* added method to compute scores for an individual team participant

* add check for for case where team is not found in signed up table

* fix test in lib/scorer_spec.rb

* add test for new participant score function
