+++ .travis.yml
@@ -7,6 +7,7 @@ env:
     - TESTFOLDER=models
     - TESTFOLDER=controllers
     - TESTFOLDER=helpers
+    - TESTFOLDER=lib
 language: ruby
 dist: trusty
 cache:
