+++ app/models/participant.rb
@@ -1,4 +1,5 @@
 class Participant < ActiveRecord::Base
+  include Scoring
   has_paper_trail
   belongs_to :user
   belongs_to :topic, class_name: 'SignUpTopic', inverse_of: false
@@ -93,7 +94,14 @@ class Participant < ActiveRecord::Base
 
   # Return scores that this participant for the given questions
   # Implemented in assignment_participant.rb
-  def scores(questions); end
+  def scores(questions) 
+    scores = {}
+    scores[:participant] = self 
+    ResponseMap.compute_assignment_score(self, questions, scores)
+    scores[:total_score] = compute_total_score(self.assignment, scores)
+    ResponseMap.merge_scores(self, scores) if self.assignment.varying_rubrics_by_round?
+    scores
+  end
 
   # Authorizations are paricipant, reader, reviewer, submitter (They are not store in Participant table.)
   # Permissions are can_submit, can_review, can_take_quiz.
