+++ config/application.rb
@@ -28,7 +28,8 @@ module Expertiza
     # Configure sensitive parameters which will be filtered from the log file.
     config.filter_parameters += [:password, :password_confirmation, :password, :password_confirmation]
     config.active_record.whitelist_attributes = false
-    config.autoload_paths += Dir[Rails.root.join('lib', '{**}')]
+    config.autoload_paths << Rails.root.join('lib', '{**}')
+    config.eager_load_paths << Rails.root.join('lib')
     config.react.addons = true
     config.assets.paths << Rails.root.join('vendor', 'assets', 'components')
     config.cache_store = :redis_store, "redis://#{ENV.fetch('REDIS_HOST', 'localhost')}:6379/0/cache", { raise_errors: false }
