+++ spec/models/assignment_team_spec.rb
@@ -390,7 +390,7 @@ describe 'AssignmentTeam' do
         allow(Response).to receive(:compute_scores).with(scores[:questionnaire2][:assessments], questions[:questionnaire2]).and_return(5)
         allow(questionnaire1).to receive(:symbol).with(no_args).and_return(:questionnaire1)
         allow(questionnaire2).to receive(:symbol).with(no_args).and_return(:questionnaire2)
-        allow(team.assignment).to receive(:compute_total_score).with(scores.except(:total_score)).and_return(10)
+        allow(team).to receive(:compute_total_score).with(team.assignment, scores.except(:total_score)).and_return(10)
         expect(team.scores(questions)).to eq(scores)
       end
     end
